[Gepardec/mega#702] separate business logic
